+++ web/src/components/ConsentBanner.js
@@ -2,18 +2,6 @@ import { Button } from '@cmsgov/design-system';
 import PropTypes from 'prop-types';
 import React, { Fragment, useState } from 'react';
 
-const cookie = name => {
-  const cookieMap = (document.cookie || '').split(';').reduce((c, s) => {
-    const bits = s.trim().split('=');
-    if (bits.length === 2) {
-      return { ...c, [bits[0].trim()]: bits[1].trim() };
-    }
-    return c;
-  }, {});
-
-  return cookieMap[name];
-};
-
 const ConsentBanner = ({ onAgree }) => {
   const [showDetails, setShowDetails] = useState(false);
 
@@ -25,12 +13,6 @@ const ConsentBanner = ({ onAgree }) => {
   const expandDetails = () => {
     setShowDetails(true);
   };
-
-  const hasConsented = cookie('gov.cms.eapd.hasConsented');
-  if (hasConsented) {
-    agreeAndContinue();
-    return null;
-  }
 
   return (
     <main id="start-main-content">