+++ web/src/components/LoginLocked.js
@@ -1,6 +1,8 @@
 import React from 'react';
+import PropType from 'prop-types';
+import { Button } from '@cmsgov/design-system';
 
-const LoginLocked = () => {
+const LoginLocked = ({ onCancel }) => {
   return (
     <div id="start-main-content">
       <div className="ds-l-container">
@@ -20,14 +22,18 @@ const LoginLocked = () => {
             <span className="ds-u-color--error">60 minutes 00 seconds.</span>
           </p>
           <div className="ds-u-display--flex ds-u-justify-content--end ds-u-margin-top--3 ds-u-padding-top--2 ds-u-border-top--2">
-            <a href="/" className="ds-c-button ds-c-button--transparent">
+            <Button variation="transparent" type="button" onClick={onCancel}>
               Cancel
-            </a>
+            </Button>
           </div>
         </div>
       </div>
     </div>
   );
+};
+
+LoginLocked.propTypes = {
+  onCancel: PropType.func.isRequired
 };
 
 export default LoginLocked;