+++ web/src/containers/StateAccessRequest.js
@@ -5,7 +5,12 @@ import { Dropdown } from '@cmsgov/design-system';
 import AuthenticationForm from '../components/AuthenticationForm';
 import { usStatesDropdownOptions } from '../util/states';
 
-const StateAccessRequest = ({ action, errorMessage, fetching }) => {
+const StateAccessRequest = ({
+  saveAction,
+  cancelAction,
+  errorMessage,
+  fetching
+}) => {
   const [selectedStates, setStates] = useState([
     usStatesDropdownOptions[0].value
   ]);
@@ -16,7 +21,12 @@ const StateAccessRequest = ({ action, errorMessage, fetching }) => {
 
   const handleSubmit = e => {
     e.preventDefault();
-    action(selectedStates);
+    saveAction(selectedStates);
+  };
+
+  const handleCancel = e => {
+    e.preventDefault();
+    cancelAction();
   };
 
   return (
@@ -32,6 +42,7 @@ const StateAccessRequest = ({ action, errorMessage, fetching }) => {
         working={fetching}
         primaryButtonText={['Submit', 'Submitting']}
         onSave={handleSubmit}
+        onCancel={handleCancel}
       >
         <div className="ds-u-margin-bottom--4">
           <label
@@ -47,7 +58,7 @@ const StateAccessRequest = ({ action, errorMessage, fetching }) => {
             name="states"
             options={usStatesDropdownOptions}
             size="medium"
-            value={selectedStates}
+            value={selectedStates[0]}
             onChange={changeStates}
           />
         </div>
@@ -57,9 +68,10 @@ const StateAccessRequest = ({ action, errorMessage, fetching }) => {
 };
 
 StateAccessRequest.propTypes = {
-  errorMessage: PropTypes.bool,
+  errorMessage: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
   fetching: PropTypes.bool.isRequired,
-  action: PropTypes.func.isRequired
+  saveAction: PropTypes.func.isRequired,
+  cancelAction: PropTypes.func.isRequired
 };
 
 StateAccessRequest.defaultProps = {