+++ web/src/containers/StateAccessRequest.test.js
@@ -8,8 +8,9 @@ let renderUtils;
 describe('<StateAccessRequest />', () => {
   beforeEach(() => {
     props = {
-      errorMessage: false,
-      action: jest.fn(),
+      errorMessage: null,
+      saveAction: jest.fn(),
+      cancelAction: jest.fn(),
       fetching: false
     };
     renderUtils = renderWithConnection(<StateAccessRequest {...props} />);
@@ -23,7 +24,7 @@ describe('<StateAccessRequest />', () => {
   it('allows the user to select the first item', () => {
     const { getByText } = renderUtils;
     fireEvent.click(getByText(/Submit/i));
-    expect(props.action).toHaveBeenCalledWith(['al']);
+    expect(props.saveAction).toHaveBeenCalledWith(['al']);
   });
 
   it('allows the users to select any item', () => {
@@ -32,6 +33,12 @@ describe('<StateAccessRequest />', () => {
       target: { value: 'mo' }
     });
     fireEvent.click(getByText(/Submit/i));
-    expect(props.action).toHaveBeenCalledWith(['mo']);
+    expect(props.saveAction).toHaveBeenCalledWith(['mo']);
+  });
+
+  it('allows the user to cancel the request', () => {
+    const { getByText } = renderUtils;
+    fireEvent.click(getByText(/Cancel/i));
+    expect(props.cancelAction).toHaveBeenCalled();
   });
 });